requirements.txt
#pip install torch==2.0.1+cu118 torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#pip install -r ./requirements.txt

clip @ git+https://github.com/openai/CLIP.git
einops
numpy~=1.26.4
open-clip-torch
opencv-python==4.6.0.66
pillow~=10.3.0
scipy==1.11.1
timm
tokenizers
torch==2.0.1+cu118
torchaudio
torchdata==0.6.1
torchmetrics
torchvision~=0.15.2
tqdm~=4.66.4
transformers==4.35.2
#triton==2.0.0
urllib3
xformers
streamlit-keyup==0.2.0
lpips~=0.1.4
clean-fid
peft~=0.11.1
dominate
diffusers==0.25.1
gradio==3.43.1
requests~=2.31.0
matplotlib~=3.9.0
clip~=1.0
accelerate~=0.30.1